Updated installation instructions, changed np.int64 to np.integer in TaskInput for compatibility with Windows

+++ pynpoint/util/multiproc.py
@@ -29,7 +29,7 @@ class TaskInput:
 
     @typechecked
     def __init__(self,
-                 input_data: Union[np.ndarray, np.int64, tuple],
+                 input_data: Union[np.ndarray, np.integer, tuple],
                  job_parameter: tuple) -> None:
         """
         Parameters
